Implementation Object Collision Estimator Library
This library provides the core functionality for the Object Collision
Estimator.

The collision estimator takes a list of objects detected by the
perception stack and the local path produced by the local planner as
inputs. It then predicts any collisions between the ego vehicle and
static obstacles. Finally the local path is modified to avoid any
collisions.

Smoothing is also done on the velocity profile of the modified path by
using a gaussian filter. This is done in order to provide a more
realistically executable local path.

Change-Id: Ib94985fda81d6dc939aa213f2711cdddd96ab390
Signed-off-by: Liyou Zhou <liyou.zhou@arm.com>
Issue-Id: SCM-1123