Remove the points too close from the origin in ray_ground_classifier
If the lidar gives too many points close to the origin, they all get
in the bin 0, effectively overflowing it. This fix just skip the points
too close to the origin. Those points weren't making much sense because
of floating point error considerations. radius < 10e-6 m. The tests
has been modified to adapt to this change.

Change-Id: Ie992fc378602fbb18d250b24ff94365ca511e8bf
Issue-ID: SCM-1007
Signed-off-by: Yoan Picchi <yoan.picchi@arm.com>